Change datatype for price to float

+++ migrations/20210106163424-modify-products.js
@@ -11,7 +11,7 @@ module.exports = {
     await queryInterface.addColumn(
       "products",
       "price",
-      { type: Sequelize.INTEGER },
+      { type: Sequelize.FLOAT },
       {}
     );
   },
